Move formatter to more accessible location
